requirements.txt
# Patient Appointment Prediction - Complete Requirements

# Core data science libraries
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit==1.39.0
# Machine learning libraries
scikit-learn>=1.0.0
xgboost>=1.5.0
shap

# Model persistence
joblib>=1.1.0

# Environment configuration
python-dotenv>=0.19.0

# Required dependencies for pandas
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7

# Required dependencies for matplotlib
pillow>=8.0.0
cycler>=0.10
pyparsing>=2.3.1
packaging>=20.0

# Required dependencies for scikit-learn
scipy>=1.6.0
threadpoolctl>=3.1.0

# Required dependencies for xgboost
scipy>=1.6.0
